refactor: rename types.internal.ts to _types.ts
• Move types.internal.ts to _types.ts for better file organization

• Update imports in calculate-cost.ts, data-loader.ts, and mcp.ts
